+++ src/Thumbnails/index.tsx
@@ -1,11 +1,12 @@
-import * as React from 'react'
+import React, { forwardRef, Children, isValidElement, cloneElement } from 'react'
 import { Box } from '@mui/material'
 
 // Types
+import type { ForwardedRef, ReactElement } from 'react'
 import type { SpinoramaThumbnailsProps } from './index.types'
 import type { SpinoramaThumbnailProps } from '../Thumbnail/index.types'
 
-const SpinoramaThumbnails: React.FC<SpinoramaThumbnailsProps> = (props: SpinoramaThumbnailsProps) => {
+const SpinoramaThumbnails = forwardRef((props: SpinoramaThumbnailsProps, ref: ForwardedRef<SpinoramaThumbnailsProps>) => {
 	// Props
 	const { justify, reverse, thumbnailsref, className, children } = props
 
@@ -13,7 +14,7 @@ const SpinoramaThumbnails: React.FC<SpinoramaThumbnailsProps> = (props: Spinoram
 	const justifyContent = !justify || justify === 'center' ? 'center' : justify === 'end' ? 'flex-end' : 'flex-start'
 
 	return (
-		<Box display="flex" flexGrow={1} width={1} {...props} className={`spinorama-thumbnails${className ? ` ${className}` : ''}`}>
+		<Box ref={ref} display="flex" flexGrow={1} width={1} {...props} className={`spinorama-thumbnails${className ? ` ${className}` : ''}`}>
 			<Box
 				className="spinorama-thumbnails-wrapper"
 				flexDirection={reverse ? 'row-reverse' : 'row'}
@@ -22,22 +23,22 @@ const SpinoramaThumbnails: React.FC<SpinoramaThumbnailsProps> = (props: Spinoram
 				minWidth={1}
 				overflow="hidden"
 			>
-				{React.Children.map(children, (child, index) => {
-					if (React.isValidElement(child)) {
+				{Children.map(children, (child, index) => {
+					if (isValidElement(child)) {
 						// Type
 						const childType = child.type.toString()
 
 						// Clone element
 						if (childType.indexOf('spinorama-thumbnail') > -1) {
-							return React.cloneElement(child as React.ReactElement<SpinoramaThumbnailProps>, {
+							return cloneElement(child as ReactElement<SpinoramaThumbnailProps>, {
 								thumbnailsref: thumbnailsref
 							})
-						} else return React.cloneElement(child)
+						} else return cloneElement(child)
 					} else return child
 				})}
 			</Box>
 		</Box>
 	)
-}
+})
 
 export default SpinoramaThumbnails