+++ src/Wrapper/index.tsx
@@ -1,14 +1,15 @@
-import * as React from 'react'
+import React, { lazy, forwardRef, Children, isValidElement, cloneElement } from 'react'
 import { Box, Grid } from '@mui/material'
 
 // Types
+import type { ForwardedRef, ReactElement } from 'react'
 import type { SpinoramaWrapperProps } from './index.types'
 import type { SpinoramaItemProps } from '../Item/index.types'
 
 // Components
-const SpinoramaItem = React.lazy(() => import('../Item'))
+const SpinoramaItem = lazy(() => import('../Item'))
 
-const SpinoramaWrapper: React.FC<SpinoramaWrapperProps> = (props: SpinoramaWrapperProps) => {
+const SpinoramaWrapper = forwardRef((props: SpinoramaWrapperProps, ref: ForwardedRef<SpinoramaWrapperProps>) => {
 	// Props
 	const { justify, reverse, className, children } = props
 
@@ -16,7 +17,7 @@ const SpinoramaWrapper: React.FC<SpinoramaWrapperProps> = (props: SpinoramaWrapp
 	const justifyContent = justify === 'end' ? 'flex-end' : 'flex-start'
 
 	return (
-		<Box display="flex" width={1} overflow="hidden" {...props} className={`spinorama-wrapper${className ? ` ${className}` : ''}`}>
+		<Box ref={ref} display="flex" width={1} overflow="hidden" {...props} className={`spinorama-wrapper${className ? ` ${className}` : ''}`}>
 			<Grid
 				className="spinorama-items"
 				container
@@ -27,16 +28,16 @@ const SpinoramaWrapper: React.FC<SpinoramaWrapperProps> = (props: SpinoramaWrapp
 				minWidth={1}
 				height="100%"
 			>
-				{React.Children.map(children, (child, index) => {
-					if (React.isValidElement(child)) {
+				{Children.map(children, (child, index) => {
+					if (isValidElement(child)) {
 						// Clone element
-						if (child.type.toString().indexOf('spinorama-item') > -1) return React.cloneElement(child as React.ReactElement<SpinoramaItemProps>, {})
+						if (child.type.toString().indexOf('spinorama-item') > -1) return cloneElement(child as ReactElement<SpinoramaItemProps>, {})
 						else return <SpinoramaItem key={index}>{child}</SpinoramaItem>
 					} else return child
 				})}
 			</Grid>
 		</Box>
 	)
-}
+})
 
 export default SpinoramaWrapper